requirements.txt
requests
beautifulsoup4
openai
reportlab
IPython
python-dotenv
streamlit
langchain_groq
langchain_huggingface
langchain
langchain_core
langchain_community
faiss-cpu
sentence-transformers
fpdf
PyPDF2
pypdf
unstructured
transformers
